+++ client/src/components/HomeGuestBase.tsx
