+++ client/src/views/HomeLogin.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import HomeGuest from ".";
-import Form from "../Form";
-import Input from "../Form/Input";
-import Button from "../Navbar/Button";
+import HomeGuest from "../components/HomeGuestBase";
+import Form from "../components/Form";
+import Input from "../components/Input";
+import Button from "../components/Button";
 
 export default function HomeLogin({linkClick}: {linkClick: () => void}) {
   return (
